middleware: Add check-env rule + check-env.sh script
Since we are using the "vendor method" of managing 3rd party
dependencies, we need to ensure that the repository is checked
out under ${GOPATH}/src.

For more information, see:
https://github.com/golang/go/wiki/GOPATH#vendor-directories-ignored-outside-of-gopath

This commit just verifies that the repo is checked out under
`${GOPATH}/src`, and prints a friendlier error message in case
it isn't.

It is indeed a bit complicated / messy to have to enforce such
restrictions on the users' environment to allow the build to
succeed, and moreso since the Go ecosystem is also partially
through a migration away from `GOPATH` (towards "Go
modules") at time of writing..

After this commit, users who want to do the Go build for
`middleware/` will at least get clear instructions on what's
going wrong if their environment is not ready for the build.

Next step might be either:
a. dockerizing the Go builds, or otherwise providing a known good
environment so user's who would like to build an Armbian image don't
need the Go toolchain + configuration at all
b. switching to using "Go modules", which doesn't require any specific
`${GOPATH}` structure, if we can assume that Go 1.11+ is available in
the build environment [1]

[1] See [this link](https://github.com/golang/go/wiki/Modules#how-do-i-use-vendoring-with-modules-is-vendoring-going-away)
if you are interested in the details)