Merge middleware repo into bitbox-base (#28)
* middleware: Initial commit

This adds the code from the separate base-middleware repo as of:
https://github.com/shiftdevices/base-middleware/commit/6e0750549

* middleware: Adjust Makefile to default to aarch64 build

* tools/bbbfancontrol: Make aarch64 build the default rule

Also add `copy-service` rule, and produce the build outputs
to top-level `build/` directory.

This setup will simplify the Armbian build process, by making
the built binaries and `.service` files available in the
`build/` directory.

* tools: Add tools-level Makefile

This just calls the `bbbfancontrol/Makefile` build rules
for now, and simplifies the build process of binaries
used as input to the Armbian image build process.

* tools/bbbfancontrol: Allow Makefile to be called from any directory

This is needed if calling this Makefile from another one,
such as the one in `tools/`, since the relative references
won't end up where the rules expect them to.

* middleware: Allow Makefile to be called from any directory

This is needed if calling this Makefile from another one,
since the relative references won't end up where the rules
expect them to.

* Makefile: Add top-level Makefile

This allows a single invocation to build both the Go
binaries (`tools/` and `middleware/`, currently) and
the Armbian image that uses those binaries as inputs.

* armbian: Allow Makefile to be called from any directory

This is needed if calling this Makefile from another one,
since the relative references won't end up where the rules
expect them to.

* armbian: Copy in prebuilt Go binaries during image build

By doing the build on the host instead of inside the Armbian build
VM, we simplify the `customize-armbian-rockpro64.sh` script and
the overall build process.

* middleware: Add check-env rule + check-env.sh script

Since we are using the "vendor method" of managing 3rd party
dependencies, we need to ensure that the repository is checked
out under ${GOPATH}/src.

For more information, see:
https://github.com/golang/go/wiki/GOPATH#vendor-directories-ignored-outside-of-gopath

This commit just verifies that the repo is checked out under
`${GOPATH}/src`, and prints a friendlier error message in case
it isn't.

It is indeed a bit complicated / messy to have to enforce such
restrictions on the users' environment to allow the build to
succeed, and moreso since the Go ecosystem is also partially
through a migration away from `GOPATH` (towards "Go
modules") at time of writing..

After this commit, users who want to do the Go build for
`middleware/` will at least get clear instructions on what's
going wrong if their environment is not ready for the build.

Next step might be either:
a. dockerizing the Go builds, or otherwise providing a known good
environment so user's who would like to build an Armbian image don't
need the Go toolchain + configuration at all
b. switching to using "Go modules", which doesn't require any specific
`${GOPATH}` structure, if we can assume that Go 1.11+ is available in
the build environment [1]

[1] See [this link](https://github.com/golang/go/wiki/Modules#how-do-i-use-vendoring-with-modules-is-vendoring-going-away)
if you are interested in the details)