Define rough architecture for middleware
The functionality of the middleware is now split into three packages. The
middleware package takes care of communicating with bitcoind and
lightningd and emits an event if new data is received.

The handlers package takes care of the communication with
the bitbox-wallet-app. It upgrades connections to websocket, if
requested, and starts the main middleware event loop. Events from the
middleware are caught and passed into a websocket handler.

The main package is what is compiled into a binary. It parses command
line arguments and creates new middleware and handler instances.
This commit also implements logging, where each package creates its own
logrus logging instance.