Fully dockerize Go builds
With this change, `make docker-build` at the top of the repo
will now perform the build of the tools/ and middleware/
inside a Docker container. The repository's root is mounted
at the expected location for the Go toolchain inside the container
(`$GOPATH/src/github.com/digitalbitbox/bitbox-base`). The build
outputs are copied to `build/` at the end of the process, and
are persisted to the host due to the volume mount.

A user which has the Go toolchain installed on their system
can still do e.g `cd middleware && make` to build directly on their
system.

Also removes middleware/scripts/ci.sh and instead adds the
contents of the script at the `make ci` target for middleware/.

Also adds the `mkdir -p $(REPO_ROOT)/build` command consistently
as a `make build-target-exists` target for all Makefiles which
need it.

Also breaks out a `envinit.sh` script that fetches the `golangci-lint`
and `dep` tools.