Makefile: Use the dockerized build for Go tools by default
Before this change, there was some issue with dependencies
with the default `make` target for the Go tools.

Specifically, in `middleware/main.go` we import other packages:

```
github.com/digitalbitbox/bitbox-base/middleware/src
github.com/digitalbitbox/bitbox-base/middleware/src/handlers
```

The user who has checked out their own fork of the `bitbox-base`
repo might or might not have these packages under the expected path:

```
$GOPATH/src/github.com/digitalbitbox/bitbox-base/middleware/src/{,handlers}
```

A further improvement would be to either `go get` the dependencies we
need, or fail with a better error message, if we can detect this situation
reliably.

Since we have a dockerized build for the Go tools, let's use
that for the default `make` target. I've added a separate
`make build-go` that does the build on the host system (not
using Docker) as well.