Implement noise encryption on the base side.
Add a 'noisemanager' package that handles the noise setup, mutual
verification between a client and the base and writing the keys to a
file.
To cleanly handle the encryption and decryption of incoming and outgoing
websocket messages, websocket listen and read loops are implemented that
automatically encrypt and decrypt traffic. The rest of the code can
interact with the loops through the use of channels.