armbian: Make read_file() handle multiple lines in prometheus_base.py
If there was more than one line in the files being read, the
single `f.readline()` call would only return the first line. Now
that we use `f.read()`, we make sure that all lines of the file
are read. (This wasn't noticed since the files we were reading
to date only had a single line of content.)