build config BASE_MINIMAL
Add build option BASE_MINIMAL (default: true) to remove unnecessary packages.
Packages are removed in sequentially in a loop to make it more resiliant if
packages are not installed, but this takes time (~3 mins) so it can be turned off.

Packages are also not removed when running the script directly on the device.