Add bitcoin resync system call
The location of the config script that runs the resync can be passed to the middleware binary as a command line argument. To facilitate the passing of these command line arguments, the arguments are now first written to a map and then read out of the map again in the system package.