make c-lightning & electrs wait for bitcoind IBD
The systemd services for c-lightning and elects check for the existence
of the /data/systemd-triggers/bitcoind-fully-synced file. If it is not
present, the unit fails on start and tries again 30s later.

Extends the 'bbb-config exec bitcoin_reindex' command to also
delete the electrs indexes and delete the trigger file.