middleware: Implement GOB RPC server
Each websocket connection now spawns a new gob encoded rpc server.
Protobuf encoding is dropped for this purpose.

The rpc server package creates read and write channels that are wrapped
into a io.ReadWriteCloser interface. This interface is implemented in
the rpcConn struct of the server package. The websocket loop in the
handlers package now uses the channels to read and write websocket
messages.

When a new websocket connection is created in the wshandler in the
handlers package, a new rpc server and subsequent websocket read and
write loop is spawned. The rpcserver's write channel is also passed into
a map of all clients and can then receive notification events again
through the listenEvents() function.

The available rpc calls are implemented in the rpcserver.go file.
Currently they are implemented as methods of RPCServer as GetSystemEnv(),
ResyncBitcoin() and GetSampleInfo(). By passing the RPCServer struct
to the served rpc connection these methods are implicitly registered.

Each of these RPCServer methods call a subsequent Middleware method
and return a response struct in each case.