General stability and reliability improvements (#117)
* rebase to pr/111

* install overlayroot package for Ubuntu build

* dashboard: add TMPFS measures

* set console=display in armbianEnv.txt

This u-boot configuration disables the serial console ttyS2 over UART and uses only the display. Prerequisite to use the BitBox secure platform over this UART port.

* disable Armbian ramlog if overlayroot is enabled

* remove custom MOTD

remove the custom BitBox Base 'message of the day' as this is hard to maintain in the read-only filesystem and is not really necessary.

* add 'bbb-config set bitcoin_dbcache' method

* remove duplicate chown in startup script

* add stresstest-ssd.sh

* remove merge conflicts

* configure logrotate service & timer

* update Prometheus and Node Exporter

fixes the issue that two hours of metrics are lost on reboot, see
https://github.com/prometheus/tsdb/issues/590

* fix swapfile setup on boot

* configure systemd journal configuration and persistant storage

* make c-lightning & electrs wait for bitcoind IBD

The systemd services for c-lightning and elects check for the existence
of the /data/systemd-triggers/bitcoind-fully-synced file. If it is not
present, the unit fails on start and tries again 30s later.

Extends the 'bbb-config exec bitcoin_reindex' command to also
delete the electrs indexes and delete the trigger file.

* set conservative dbcache 300M

The bbbsupervisor increases the bitcoind dbcache to 2GB during initial sync.
In case of a reboot, the dbcache is set back to the conservative 300MB by
default until the bbbsupervisor increases it again.

* log bitcoind restarts in .bitcoin/restarts.log

* stop lightningd / electrs on bitcoind resync

* turn off NGINX access log & limit logsize

* add error msg to electrs startup

systemd-electrs-startchecks.sh replaces individual ExecStartPre checks to provide
better error message in the startup log instead of failing without comment.

* variable prometheus-bitcoind query delay

* rename dir /systemd-triggers to /triggers

* create /data_source on build

standard build links from /data to /data_source, but Mender build mounts /data as own partition, data needs to be copied on first boot

* move /opt/shift/sysconfig  to /data/sysconfig

* remove symlink to /data

Mender build fails when it cannot create mountpoint due to existing symlink.
For non-Mender images, the symlink from /data to /data_source is created
by the startup script on first boot.

* consistently name trigger bitcoind_fully_synced

* create /data directory, link /etc/hostname and ssl key/certs

* create /data_source with symlink, delete link at end of build

standard build links from /data to /data_source on first boot, but
Mender build mounts /data as own partition, data needs to be copied on first boot

create symlink for all scripts to work, remove it at the end of build process

* remove maxmempool limitation

limited mempool can lead to transaction being dropped from it, skewing
fee estimation and mempoool statistics.

* add `PartOf=bitcoind.service` to lightningd and electrs unit

making these two services part of the bitcoind service causes them
to restart if bitcoind is restarted. This is important for the
cookie authentication to propagate correctly.

* always recreate .cookie.env before starting electrs

* add method 'bbb-config.sh set bitcoin_ibd <true|false>'