Add the SetHostname and GetHostname RPCs
The SetHostname RPC restricts what hostnames can be set.
The restrictions are a subset of what the linux man pages define as a
valid hostname (http://man7.org/linux/man-pages/man7/hostname.7.html).

The Base hostname must be:
- At least two chars long and not longer than 64 chars
- All chars must be lowercase `a-z`, `0-9`. Dots and hypens (-) are,
allowed, but can't be the first first or last chars

To enforce the character restrictions the following regex is used in
combination with a check if the hostname is at least two chars.

   `^[a-z0-9]+[a-z0-9-.]{0,62}[a-z0-9]$`