Add dummy UserAuthenticate +UserChangePassword RPC
dummy UserAuthenticate:
The dummy RPC lets a user authenticate with "admin/CanIHasPassword" as
long as the Base is not set up yet. No other user/password combination
is valid. Once the Base is set up (in the dummy by changing the
admin password via `UserChangePassword`) valid logins are:
- "admin/<user choosen admin password>"
- "satoshi/shift1"
- "dev/dev"

dummy UserChangePassword:
The dummy RPC is only successful, if the password is longer than 7
chars. Only the password for the "admin" user is changed, all other
passwords stay the same. For this dummy the Base gets initialized,
when the admin user password is changed and the Base is not
initialized.