bbbconfgen: enable #checkTrue or #checkFalse
The #check function that drops a line if the referenced Redis
key is empty or false, is extended to also accept the values
'false', 'no' or 'n' (case insensitive). It's aliased with #checkTrue.

The new inverse function #checkFalse drops the line if
the corresponding Redis value is '1', 'true', 'yes' or 'y'
(case insensitive).