bitcoind-rpcauth: script and service configuration (#179)
* redis.sh.inc: does not abort if Redis not available, unless redis_require()

* bitcoind-rpcauth: script and service configuration

Add rpcauth.py script from bitcoin repository to generate unique but persistent
rpcauth credentials for bitcoind. The script is extended to store the credentials
directly in Redis.

To ensure unique credentials, the bitcoind-startpre script calls the command
'bbb-cmd.sh bitcoind refresh_rpcauth' if no valid rpcauth value is found in Redis
(min length of 90 characters), or bitcoind:refresh-rpcauth is set to 1.

Configuration of bitcoind, lightningd, elects and bbbmiddleware is adjusted to
use new authentication method:
* bitcoind.conf contains rpcauth setting: rpcuser plus salted/hashed password
* other services use rpcuser/rpcpassword to authenticate themselves

File access to config files containing rpcpassword is restricted to service
user and group bitcoin. Two improvements should still be considered:
* storing the credentials in unsecured Redis db is not optimal
* Electrs should support rpcuser/rpcpassword without passing them through
  the command line (password visible in ps output)

As bitcoind does not manage funds, local rpcaccess (available only if system
is already compromised) is not really critical and this solution good enough
for an MVP.

* cleanup Makefile

* consistently use chmod u/g/o syntax