Bash scripting foundation for Mender over-the-air updates.
The current firmware version of the build is hardcoded into the textfile
/opt/shift/config/version and is not meant to be changed. On boot, it is
written into Redis key 'base:version' to access from other programs.

The update process is initiated by the Middleware (not part of this commit)
by calling

  bbb-cmd.sh mender-update install <version>

This script checks the version number and calls 'mender -update' with a
constructed GitHub releases download link
(e.g. https://github.com/digitalbitbox/bitbox-base/releases/download/0.0.2/BitBoxBase-v0.0.2-RockPro64.base)

Redis key 'base:updating' is used to keep track of the update progress. After
the update is applied, it is set to 10 and the Middleware restarts the Base.

Valid status codes of 'base:updating'
   0: no update in progress
  10: mender update applied
  20: system reconfigured
  30: system tested OK
  40: mender update committed
  90: mender update failed

On every boot, the script systemd-update-checks.sh (running as oneshot unit
update-checks.service) checks if an update is in progress and executes the
following tasks:

* configuration files are recreated and relevant services are restarted --> code 20

* the system is verified, all services must be up and running within 3 minutes --> code 30
  >> bbb-systemctl.sh verify

* on success, the update is committed --> code 40
  >> bbb-cmd.sh mender-update commit

* on failure, a fallback reboot is initiated --> code 50
  >> bbb-cmd.sh base restart

* if update process is finished, reset 'base:updating' to 0