bitcoind: add 'sysperms=1' to config
This setting instructs bitcoind to create new files (e.g. blocks and indices)
with system default permissions, instead of umask 077 (-rwx------).

It has been removed in error, but is needed for members of the bitcoin group
like electrs to have access to the blockchain data.

Access Control Lists (ACL) have also been added to control the default
permissions (set on startup with 'setfacl').