refactor: ErrorResponse Error() method
The logging output for a print ErrorResponse was previously
uninformative and not well formatted. The Error() method is changed
to print just `ErrorResponse: Success: true` when there is no Error.
If an Error occurs the Success, Code and Message values printed in
multiple lines. This should increased readability in the logs.