change: Restart & ShutdownBase without bbb-cmd.sh
The RestartBase RPC and the ShutdownBase RPC should both return a
response to the calling App before the shutdown is started. The
previously used os.exec's `CombineOutput` waited for the script to
complete and forking in the shell script did not resolve this.

The RPCs now execute the reboot/shutdown with a delay of 5 seconds.
The middlware responds to the App prior to the shutdown/reboot.