Middleware: Add Password Authentication with JWT and Bcrypt
RPC calls to the middleware need to be authenticated after this pull request
is merged.  To authenticate, a client first calls the UserAuthenticate method.
On first setup of the base, the username and password are set to 'admin' and
'ICanHasPasword?'. The passed in password is then compared with a bcrypt hash
of the password in redis. Redis stores a JSON 'auth' structure for this that
can be accessed with the 'auth' key. The structure maps between a username
and its password and role.

Once the user's password is successfully compared he is issued a JWT (JSON
Web Token). He then needs to include this token in every rpc call to be
successfull. The rpc methods return an error otherwise.

To change the password, the UserChangePassword method first checks the length
of the new password. Then it fetches the 'auth' structure from redis, checks
if the username exists and the provided current password is correct. Once done
the new password is hashed and encoded with bcrypt and finally saved to redis.

A lot of unit tests had to be adapted for this change. Most importantly the rpc
tests now only test against a mocked version of the middleware, which was
generated with mockery.

I removed some units tests from the rpcserver test that I thought rather test
the middleware's behaviour, than the rpcserver's.

To control the flow of the setup wizard, the middleware also tracks if
the user set the password yet and if the lightning backup was created
and the user subsequently completed setup. These variables are tracked
in redis.