add: first version of GetServiceInfo RPC
This commit adds a first version of the GetServiceInfo RPC.
The RPC returns information about services such as bitcoind, electrs,
and lightningd. Information includes for example the current block
height or the count of connected peers.
The App should call this RPC once it receives a notification that the
serviceInfo in the middleware has changed.