add: Base update RPCs
The BitBoxBase should support over-the-air firmware updates. This
commit adds the two necessary RPCs for this.

UpdateBase RPC:
This RPC performs an update of the Base firmware to a specified
version. This includes downloading, applying and rebooting. Every
time the download progress or the state changes a notification is
send to the App.

GetBaseUpdateProgress RPC:
This RPC should be called when the app receives a notification about
a changed state. It returns the state, downloaded byte and progress
in percent.