refactor: improve get flashdrive name
This commit improves the response parsing of the flashdrive check
script call. On a responses with multiple lines a warning is logged.
On a response with no lines returns an ErrorResponse with a code of
ErrorUnexpected is returned. The flash drive name is now the last
line of the script output.