add: Base firmware update check and notification
This adds a periodic check for firmware updates to the middlware.
If an update is available, a the middleware sends a notification.
The App can get more information about the update by calling the
IsBaseUpdateAvailable RPC, which returns a boolean indicating if
an update is available and information about the update.

The App should call IsBaseUpdateAvailable after the connection is
established.