add: read IPC notifications from a named pipe
This defines IPC notifications and adds a IPC notification reader,
which creates and opens a named pipe. Other processes, for example
shell or python scripts, can write notifications into the named pipe.

Notifications are sent in a JSON encoded string. The protocol in
version 1 requires notifications to have a version, topic and a
payload. Notifications can have a maximum length of 4096-1 byte.