middleware/hsm: increase read timeout
There are blocking operations like verifying the pairing code on the
HSM, which should not timeout in 5s.

A proper fix is to have the HSM timeout itself by default after a
defined period, and have the client read timeout be larger than this.

For now we just increase the timeout to give enough time hopefully.