add: handling of state for update+shutdown+reboot
This adds the handling of the following system states to the
Supervisor:
- DOWNLOAD_UPDATE
- UPDATE_FAILED
- REBOOT
- SHUTDOWN

The Middleware logs a destinct logtag once one of these states become
active. The Supervisor watches the Middleware log and handles the
state changes by setting the respective `descriptionCodes` in the
sorted set in Redis and the respective stateCode. The Middleware is
notified that there was a state change in Redis via the named pipe
IPC, but doesn't handle the notification yet.