change: use struct for Config factory
This changes the `ǸewConfiguration` factory function to use a struct
as parameter and not the single values. This is done, because structs
allow named fields, while Go functions don't support named fields.
Not using named fields can lead to switched parameters. Named fields
help to avoid that and make code review easier.

The previous commit fixes a bug related calling the factory function
with switched parameters.