hsm-config: set hsm config on middleware start and SetHostname
This commit:
Adds the setHSMConfig() function to call the HSM API's
BitBoxBaseSetConfig() on middleware startup and when the Base hostname
is changed using an RPC.
The IP and hostname are fetched from redis and prometheus respctively
whenver setHSMConfig is called.
For statusLedMode and statusScreenMode we currently use the default
LED_ALWAYS and SCREEN_ALWAYS until we have redis config for those values.