Prometheus: log public network availability
https://github.com/shiftdevices/bitbox-base-internal/issues/370

Prometheus should log the availability of public internet without
leaking privacy information.

This pull requests queries an external host with a lot of general
traffic, Cloudflare, over Tor:

```
curl --socks5-hostname localhost:9050 1.1.1.1
```

If Tor is not active, it's impossible to ping an external host without
revealing the own ip address, but the solution can just "mingle in the
crowd", e.g. by `ping`ing Cloudflare, which should not be suspicious.

```
ping -c 1 1.1.1.1
```

These queries are run regularly from `prometheus-base.py`, so that the
result is collected by Prometheus and stored in its time-series database.
This helps with analyzing/debugging incidents, as the public network
availability can also be queried after the fact.

This commit:
* provides the new Prometheus metric 'base_internet_connectivity' that
  is 0 when OK, or an error code when NOT OK