HSM: add firmware to Base image
https://github.com/shiftdevices/bitbox-base-internal/issues/369

Currently, we update the HSM firmware together with the Base image.
For the Middleware to update the HSM, the specific version of the
HSM firmware needs to be included in the disk image.

This issue extends the build script to provide a specific release
version in the disk image. The version is specified like `
HSM_VERSION='4.3.0'`, that is then used to create the valid download
URI like this:

```
https://github.com/digitalbitbox/bitbox02-firmware/releases/download/firmware-bitboxbase%2Fv4.3.0/firmware-bitboxbase.v4.3.0.signed.bin
```

It is important, that the `tag` of the release always follows this
syntax: `firmware-bitboxbase/v4.3.0` and the file is named like
`firmware-bitboxbase.v4.3.0.signed.bin`.

The release binary file is then stored in the disk image:

```
/opt/shift/hsm/firmware-bitboxbase.signed.bin
```

The version is deliberately omited in the filename, as the Middleware
reads the internal version number (of the monotonic counter) directly
out of the file, and compares it with the currently installed firmware
on the HSM.

This commit:
* extends the build script