cleanup: RPCServer log messages for RPC replys
The RPCServer logged the same "sent reply" message for each RPC
response. The output contained values, but no field labels for the
values.

This adds the name of the RPC to the response log and prints the
responses with field names. This shoudl make the Middleware logs
more accessible for debugging.