lightingd-startpost: check of Socket, not file
The script that is started by systemd after ligntingd checks if the
'lightning-rpc' socket (mainnet or testnet) exists, and if so sets
additional GROUP permissions (rwx).

It's a bug that it checks using bash's `-f` directive, which only
returns true for regular files. It either needs to check for
any file `-e` or specifically a socket `-S`.

This commit:
* checks with `-S` if the lightning-rpc socket exists