build: use Ayufan kernel config
Stability of the BitBoxBase is considerably higher using the Ayufan
kernel. A first step in using it is to build the kernel within the
Armbian build process with the Ayufan kernel config.

Some Armbian user patches might need to be removed still, but in
general, initial testing shows that this build works very well.

This commit:
* fixes the Armbian build repo at a (arbitrary) recent commit and
  patches the debootstrap.sh script to correctly name the Ayufan kernel
* adds the linux-rockchip64-legacy.config to userpatches